Add polls and other candidates, also catcher game

+++ app/src/main/java/com/example/qumeric/elections/JSONLoader.kt
@@ -55,7 +55,7 @@ fun loadFakeCandidates(resources: Resources): MutableList<FakeCandidate> {
         val candidate = FakeCandidate(
                 JSONcandidate.getString("name"),
                 JSONcandidate.getString("description"),
-                JSONcandidate.getInt("level")
+                JSONcandidate.getDouble("level")
         )
 
         candidates.add(candidate)
